# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OGM Gönüllü Yönetim Sistemi API",
    "version": "0.1.0",
    "description": "T.C. Orman Genel Müdürlüğü Gönüllü Yönetim Sistemi backend API'si. Auth (e-Devlet + Telefon OTP), onboarding, profil, ana sayfa, yangın bildirimi ve referans veri uçlarını içerir.",
    "contact": {
      "name": "OGM Gönüllü Backend",
      "email": "gonullu-sistem@ogm.gov.tr"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com/v1",
      "description": "API v1"
    },
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com",
      "description": "Kök (health vb.)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Erişim token'ı (`Authorization: Bearer <token>`)"
      },
      "registrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Telefon doğrulama sonrası kayıt formu için kısa süreli token"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Panel admin anahtarı (ADMIN_API_KEY)"
      },
      "officerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Saha amiri anahtarı (OFFICER_API_KEY)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "validation_error"
              },
              "message": {
                "type": "string",
                "example": "Form alanlarında hata var"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cover": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "format": "date"
          },
          "readTimeMin": {
            "type": "integer"
          },
          "themes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "author": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "format": "uri",
                "nullable": true
              }
            }
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "enum": [
                    "paragraph",
                    "heading",
                    "image"
                  ]
                },
                "text": {
                  "type": "string"
                },
                "source": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Kaynak bulunamadı",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "security": [],
  "paths": {
    "/admin/fire-reports/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Yangın bildirimi statü geçişi (panel)",
        "description": "Mobil uygulamadan ÇAĞRILMAZ. x-api-key (admin) gerektirir.",
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "confirmed",
                      "rejected"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Güncellendi"
          }
        }
      }
    },
    "/admin/missions/{id}/photos/{submissionId}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Görev fotoğrafı moderasyonu (panel)",
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "submissionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "approved",
                      "rejected"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Güncellendi"
          }
        }
      }
    },
...
    "/blog/posts": {
      "get": {
        "tags": [
          "Blog"
        ],
        "summary": "Blog yazıları",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Yazı listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPost"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/blog/posts/{id}": {
      "get": {
        "tags": [
          "Blog"
        ],
        "summary": "Blog yazısı detayı",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Yazı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "404": {
            "description": "post_not_found"
          }
        }
      }
    },
...
    "/legal/{document}": {
      "get": {
        "tags": [
          "Legal"
        ],
        "summary": "Yasal metinleri (KVKK, aydınlatma, açık rıza) versiyonlu olarak döner",
        "parameters": [
          {
            "in": "path",
            "name": "document",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "kvkk",
                "aydinlatma-metni",
                "acik-riza"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metin",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date"
                    },
                    "title": {
                      "type": "string"
                    },
                    "body": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}